dashboard.py
```python
# Import necessary libraries
import pandas as pd
import matplotlib.pyplot as plt
from google.cloud import storage

# Define constants
BUCKET_NAME = 'your-bucket-name'  # Replace with your bucket name
CSV_FILE = 'analyzed_data.csv'  # CSV file to store analyzed data

# Initialize a storage client
storage_client = storage.Client()

def download_from_bucket(blob_name, path_to_file, bucket_name):
    """Download data from a bucket"""
    bucket = storage_client.get_bucket(bucket_name)
    blob = bucket.blob(blob_name)
    blob.download_to_filename(path_to_file)

    print(f"File {blob_name} downloaded from {bucket_name}.")

def visualize_data():
    """Main function to visualize data"""
    # Download the analyzed data from the cloud storage
    download_from_bucket(CSV_FILE, CSV_FILE, BUCKET_NAME)

    # Load analyzed data from the CSV file
    df = pd.read_csv(CSV_FILE)

    # Plot a histogram of the predicted species
    df['predicted_species'].value_counts().plot(kind='bar')
    plt.title('Distribution of Animal Species')
    plt.xlabel('Species')
    plt.ylabel('Count')
    plt.show()

    # Plot a histogram of the time of day
    df['time_of_day'].value_counts().plot(kind='bar')
    plt.title('Distribution of Animal Activity by Time of Day')
    plt.xlabel('Time of Day')
    plt.ylabel('Count')
    plt.show()

if __name__ == "__main__":
    visualize_data()
```
