model_training.py
```python
# Import necessary libraries
import os
import pandas as pd
from google.cloud import storage
from sklearn.model_selection import train_test_split
from tensorflow.keras.preprocessing import image
from tensorflow.keras.applications.mobilenet_v2 import MobileNetV2, preprocess_input
from tensorflow.keras.layers import Dense, GlobalAveragePooling2D
from tensorflow.keras.models import Model
from tensorflow.keras.optimizers import Adam

# Define constants
BUCKET_NAME = 'your-bucket-name'  # Replace with your bucket name
IMAGE_DIR = 'camera_images'  # Directory where images are stored locally
CSV_FILE = 'image_data.csv'  # CSV file to store image metadata
MODEL_FILE = 'model.h5'  # File to save the trained model

# Initialize a storage client
storage_client = storage.Client()

def download_from_bucket(blob_name, path_to_file, bucket_name):
    """Download data from a bucket"""
    bucket = storage_client.get_bucket(bucket_name)
    blob = bucket.blob(blob_name)
    blob.download_to_filename(path_to_file)

    print(f"File {blob_name} downloaded from {bucket_name}.")

def load_images(image_files):
    """Load and preprocess images"""
    images = []
    for image_file in image_files:
        img = image.load_img(image_file, target_size=(224, 224))
        img_array = image.img_to_array(img)
        img_array = preprocess_input(img_array)
        images.append(img_array)
    return np.array(images)

def train_model():
    """Main function to train the model"""
    # Download the CSV file from the cloud storage
    download_from_bucket(CSV_FILE, CSV_FILE, BUCKET_NAME)

    # Load image metadata from the CSV file
    df = pd.read_csv(CSV_FILE)

    # Load and preprocess images
    X = load_images(df['image_file'])
    y = df['species']  # Assuming the CSV file contains a 'species' column

    # Split the data into training and validation sets
    X_train, X_val, y_train, y_val = train_test_split(X, y, test_size=0.2, random_state=42)

    # Load the MobileNetV2 model for transfer learning
    base_model = MobileNetV2(weights='imagenet', include_top=False)
    x = base_model.output
    x = GlobalAveragePooling2D()(x)
    predictions = Dense(len(y.unique()), activation='softmax')(x)

    # Create the final model
    model = Model(inputs=base_model.input, outputs=predictions)

    # Freeze the base model layers
    for layer in base_model.layers:
        layer.trainable = False

    # Compile the model
    model.compile(optimizer=Adam(lr=0.001), loss='categorical_crossentropy', metrics=['accuracy'])

    # Train the model
    model.fit(X_train, y_train, validation_data=(X_val, y_val), epochs=10, batch_size=32)

    # Save the trained model
    model.save(MODEL_FILE)

    # Upload the trained model to the cloud storage
    upload_to_bucket(MODEL_FILE, MODEL_FILE, BUCKET_NAME)

if __name__ == "__main__":
    train_model()
```
