requirements.txt
# Python version
python==3.8

# Libraries for data analysis
pandas==1.2.4
numpy==1.20.2

# Libraries for machine learning
scikit-learn==0.24.2

# Libraries for data visualization
matplotlib==3.4.2
seaborn==0.11.1

# Libraries for cloud storage
google-cloud-storage==1.38.0

# Libraries for creating a user-friendly dashboard
dash==1.20.0
dash-bootstrap-components==0.12.2
